Extract Interface and Misc tabs
